requirements.txt
streamlit>=1.28.0
playwright>=1.40.0
trafilatura>=1.6.0
pypdf>=4.0.0
huggingface_hub>=0.20.0
datasets>=2.16.0
python-dotenv>=1.0.0
# Entrenamiento local / RunPod (training_connection.py)
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.26.0
peft>=0.8.0
